Refactor: Moved test classes to own file
Moved all the Test classes to their own file to keep code DRY
