+++ src/test/java/com/connectfour/BoardTest.java
@@ -3,6 +3,7 @@ package com.connectfour;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+
 import static org.junit.jupiter.api.Assertions.*;
 
 class BoardTest {
@@ -14,8 +15,7 @@ class BoardTest {
     }
     @Test
     void testAddDiscPutsDiscAtBottomOfEmptyRow() {
-        // TODO: write test for addDisc;
-        Discable testDisc = new TestDisc();
+        Discable testDisc = new TestClasses.TestDisc();
         testBoard.addDisc(testDisc, 0);
 
         assertEquals( testDisc, testBoard.grid[5][0]);
@@ -24,7 +24,7 @@ class BoardTest {
 
     @Test
     void testAddDiscPlacesDiscOnTopOfOtherDiscs(){
-        Discable testDisc = new TestDisc();
+        Discable testDisc = new TestClasses.TestDisc();
         testBoard.addDisc(testDisc, 0);
         testBoard.addDisc(testDisc, 0);
 
@@ -39,22 +39,11 @@ class BoardTest {
 
     @Test
     void testFindDisc(){
-        Discable testDisc = new TestDisc();
+        Discable testDisc = new TestClasses.TestDisc();
         testBoard.grid[0][0] = testDisc;
 
         assertEquals(testDisc, testBoard.findDisc(0,0));
     }
 
-    public class TestDisc implements Discable {
 
-        @Override
-        public char getSymbol() {
-            return 'X';
-        }
-
-        @Override
-        public String getColor() {
-            return "blue";
-        }
-    }
 }
