+++ src/test/java/com/connectfour/DisplayTest.java
@@ -11,11 +11,11 @@ import static org.junit.jupiter.api.Assertions.*;
 class DisplayTest {
 
     Display display;
-    TestStream testStream;
+    TestClasses.TestStream testStream;
 
     @BeforeEach
     public void initTestStreamAndDisplay() {
-         testStream = new TestStream(System.out);
+         testStream = new TestClasses.TestStream(System.out);
          display = new Display(testStream);
     }
     @Test
@@ -48,30 +48,5 @@ class DisplayTest {
         assertEquals(message, testStream.charState);
     }
 
-    private class TestStream extends PrintStream {
-        String stringState;
-        char charState;
-        public TestStream(OutputStream out) {
-            super(out);
-        }
 
-        @Override
-        public void print(String message){
-            stringState = message;
-        }
-        @Override
-        public void println(String message){
-            stringState = message;
-        }
-
-        @Override
-        public void print(char message){
-            charState = message;
-        }
-        @Override
-        public void println(char message){
-            charState = message;
-        }
-
-    }
 }
