+++ src/test/java/com/connectfour/GameTest.java
@@ -8,10 +8,10 @@ class GameTest {
 
     @Test
     void testStartDisplaysWelcomeMessage() {
-        TestDisplay testDisplay = new TestDisplay();
-        Playerable testPlayer = new TestPlayer();
+        TestClasses.TestDisplay testDisplay = new TestClasses.TestDisplay();
+        Playerable testPlayer = new TestClasses.TestPlayer();
         Board board = new Board();
-        Scannable testScanner = new TestScanner();
+        Scannable testScanner = new TestClasses.TestScanner();
 
         Game game = new Game(testDisplay, testScanner, board, testPlayer);
 
@@ -22,81 +22,6 @@ class GameTest {
 
     @Test
     void play() {
-    }
-
-    public class TestDisplay implements Displayable {
-        String stringState;
-        char charState;
-
-        public TestDisplay(){
-            stringState = "";
-            charState = ' ';
-        }
-        @Override
-        public void print(String message) {
-            stringState = message;
-        }
-
-        @Override
-        public void println(String message) {
-            stringState = message;
-        }
-
-        @Override
-        public void print(char message) {
-            charState = message;
-        }
-
-        @Override
-        public void println(char message) {
-            charState = message;
-        }
-
-        public char getCharState(){
-            return charState;
-        }
-
-        public String getStringState(){
-            return stringState;
-        }
-    }
-
-    public class TestScanner implements Scannable {
-
-        @Override
-        public int getColumn() {
-            return 1;
-        }
-    }
-
-    public class TestPlayer implements Playerable {
-
-        @Override
-        public void takeTurn(Board board) {
-
-        }
-
-        @Override
-        public String getName() {
-            return null;
-        }
-
-        @Override
-        public Discable getDisc() {
-            return null;
-        }
-    }
-    public class TestDisc implements Discable {
-
-        @Override
-        public char getSymbol() {
-            return 'X';
-        }
-
-        @Override
-        public String getColor() {
-            return "null";
-        }
     }
 
 }
