+++ src/test/java/com/connectfour/PresenterTest.java
@@ -8,11 +8,11 @@ import java.io.PrintStream;
 import static org.junit.jupiter.api.Assertions.*;
 
 class PresenterTest {
-    TestDisplay testDisplay;
+    TestClasses.TestDisplay testDisplay;
 
     @BeforeEach
     public void initTestDisplay(){
-        testDisplay = new TestDisplay();
+        testDisplay = new TestClasses.TestDisplay();
     }
 
     @Test
@@ -27,60 +27,10 @@ class PresenterTest {
     @Test
     void testPrintDisc() {
 
-        Discable testDisc = new TestDisc();
+        Discable testDisc = new TestClasses.TestDisc();
         Presenter.printDisc(testDisplay, testDisc);
 
         assertEquals(testDisplay.getCharState(), testDisc.getSymbol());
-    }
-
-    public class TestDisplay implements Displayable {
-        String stringState;
-        char charState;
-
-        public TestDisplay(){
-            stringState = "";
-            charState = ' ';
-        }
-        @Override
-        public void print(String message) {
-            stringState = message;
-        }
-
-        @Override
-        public void println(String message) {
-            stringState = message;
-        }
-
-        @Override
-        public void print(char message) {
-            charState = message;
-        }
-
-        @Override
-        public void println(char message) {
-            charState = message;
-        }
-
-        public char getCharState(){
-            return charState;
-        }
-
-        public String getStringState(){
-            return stringState;
-        }
-    }
-
-    public class TestDisc implements Discable {
-
-        @Override
-        public char getSymbol() {
-            return 'X';
-        }
-
-        @Override
-        public String getColor() {
-            return "null";
-        }
     }
 
 }
